+++ yarn/datadog_checks/yarn/yarn.py
@@ -1,92 +1,22 @@
 # (C) Datadog, Inc. 2018
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
+import os
 
-"""
-YARN Cluster Metrics
---------------------
-yarn.metrics.appsSubmitted          The number of submitted apps
-yarn.metrics.appsCompleted          The number of completed apps
-yarn.metrics.appsPending            The number of pending apps
-yarn.metrics.appsRunning            The number of running apps
-yarn.metrics.appsFailed             The number of failed apps
-yarn.metrics.appsKilled             The number of killed apps
-yarn.metrics.reservedMB             The size of reserved memory
-yarn.metrics.availableMB            The amount of available memory
-yarn.metrics.allocatedMB            The amount of allocated memory
-yarn.metrics.totalMB                The amount of total memory
-yarn.metrics.reservedVirtualCores   The number of reserved virtual cores
-yarn.metrics.availableVirtualCores  The number of available virtual cores
-yarn.metrics.allocatedVirtualCores  The number of allocated virtual cores
-yarn.metrics.totalVirtualCores      The total number of virtual cores
-yarn.metrics.containersAllocated    The number of containers allocated
-yarn.metrics.containersReserved     The number of containers reserved
-yarn.metrics.containersPending      The number of containers pending
-yarn.metrics.totalNodes             The total number of nodes
-yarn.metrics.activeNodes            The number of active nodes
-yarn.metrics.lostNodes              The number of lost nodes
-yarn.metrics.unhealthyNodes         The number of unhealthy nodes
-yarn.metrics.decommissionedNodes    The number of decommissioned nodes
-yarn.metrics.rebootedNodes          The number of rebooted nodes
-
-YARN App Metrics
-----------------
-yarn.app.progress             The progress of the application as a percent
-yarn.app.startedTime          The time in which application started (in ms since epoch)
-yarn.app.finishedTime         The time in which the application finished (in ms since epoch)
-yarn.app.elapsedTime          The elapsed time since the application started (in ms)
-yarn.app.allocatedMB          The sum of memory in MB allocated to the applications running containers
-yarn.app.allocatedVCores      The sum of virtual cores allocated to the applications running containers
-yarn.app.runningContainers    The number of containers currently running for the application
-yarn.app.memorySeconds        The amount of memory the application has allocated (megabyte-seconds)
-yarn.app.vcoreSeconds         The amount of CPU resources the application has allocated (virtual core-seconds)
-
-YARN Node Metrics
------------------
-yarn.node.lastHealthUpdate       The last time the node reported its health (in ms since epoch)
-yarn.node.usedMemoryMB           The total amount of memory currently used on the node (in MB)
-yarn.node.availMemoryMB          The total amount of memory currently available on the node (in MB)
-yarn.node.usedVirtualCores       The total number of vCores currently used on the node
-yarn.node.availableVirtualCores  The total number of vCores available on the node
-yarn.node.numContainers          The total number of containers currently running on the node
-
-YARN Capacity Scheduler Metrics
------------------
-yarn.queue.root.maxCapacity            The configured maximum queue capacity in percentage for root queue
-yarn.queue.root.usedCapacity           The used queue capacity in percentage for root queue
-yarn.queue.root.capacity               The configured queue capacity in percentage for root queue
-yarn.queue.numPendingApplications      The number of pending applications in this queue
-yarn.queue.userAMResourceLimit.memory  The maximum memory resources a user can use for Application Masters (in MB)
-yarn.queue.userAMResourceLimit.vCores  The maximum vCpus a user can use for Application Masters
-yarn.queue.absoluteCapacity            The absolute capacity percentage this queue can use of entire cluster
-yarn.queue.userLimitFactor             The minimum user limit percent set in the configuration
-yarn.queue.userLimit                   The user limit factor set in the configuration
-yarn.queue.numApplications             The number of applications currently in the queue
-yarn.queue.usedAMResource.memory       The memory resources used for Application Masters (in MB)
-yarn.queue.usedAMResource.vCores       The vCpus used for Application Masters
-yarn.queue.absoluteUsedCapacity        The absolute used capacity percentage this queue is using of the entire cluster
-yarn.queue.resourcesUsed.memory        The total memory resources this queue is using (in MB)
-yarn.queue.resourcesUsed.vCores        The total vCpus this queue is using
-yarn.queue.AMResourceLimit.vCores      The maximum vCpus this queue can use for Application Masters
-yarn.queue.AMResourceLimit.memory      The maximum memory resources this queue can use for Application Masters (in MB)
-yarn.queue.capacity                    The configured queue capacity in percentage relative to its parent queue
-yarn.queue.numActiveApplications       The number of active applications in this queue
-yarn.queue.absoluteMaxCapacity         The absolute maximum capacity percentage this queue can use of the entire cluster
-yarn.queue.usedCapacity                The used queue capacity in percentage
-yarn.queue.numContainers               The number of containers being used
-yarn.queue.maxCapacity                 The configured maximum queue capacity in percentage relative to its parent queue
-yarn.queue.maxApplications             The maximum number of applications this queue can have
-yarn.queue.maxApplicationsPerUser      The maximum number of active applications per user this queue can have
-"""
-
-from six.moves.urllib.parse import urljoin, urlsplit, urlunsplit
 from six import iteritems
+from six.moves.urllib.parse import urljoin, urlsplit, urlunsplit
 
-from requests.exceptions import Timeout, HTTPError, InvalidURL, ConnectionError, SSLError
 import requests
+import requests_kerberos
+from requests.exceptions import Timeout, HTTPError, InvalidURL, ConnectionError, SSLError
 
-from datadog_checks.checks import AgentCheck
-from datadog_checks.config import _is_affirmative
+from datadog_checks.base import AgentCheck, is_affirmative
+
+KERBEROS_STRATEGIES = {
+    'required': requests_kerberos.REQUIRED,
+    'optional': requests_kerberos.OPTIONAL,
+    'disabled': requests_kerberos.DISABLED,
+}
 
 
 # Default settings
@@ -232,16 +162,6 @@ class YarnCheck(AgentCheck):
         # Collected by default
         app_tags['app_name'] = 'name'
 
-        # Authenticate our connection to endpoint if required
-        username = instance.get('username')
-        password = instance.get('password')
-        auth = None
-        if username is not None and password is not None:
-            auth = (username, password)
-
-        # Option to disable verifying ssl certificate
-        ssl_verify = _is_affirmative(instance.get('ssl_verify', True))
-
         # Get additional tags from the conf file
         custom_tags = instance.get('tags', [])
         tags = list(set(custom_tags))
@@ -258,17 +178,17 @@ class YarnCheck(AgentCheck):
         tags.append('cluster_name:{}'.format(cluster_name))
 
         # Get metrics from the Resource Manager
-        self._yarn_cluster_metrics(rm_address, auth, ssl_verify, tags)
-        if _is_affirmative(instance.get('collect_app_metrics', DEFAULT_COLLECT_APP_METRICS)):
-            self._yarn_app_metrics(rm_address, auth, ssl_verify, app_tags, tags)
-        self._yarn_node_metrics(rm_address, auth, ssl_verify, tags)
-        self._yarn_scheduler_metrics(rm_address, auth, ssl_verify, tags, queue_blacklist)
+        self._yarn_cluster_metrics(rm_address, instance, tags)
+        if is_affirmative(instance.get('collect_app_metrics', DEFAULT_COLLECT_APP_METRICS)):
+            self._yarn_app_metrics(rm_address, instance, app_tags, tags)
+        self._yarn_node_metrics(rm_address, instance, tags)
+        self._yarn_scheduler_metrics(rm_address, instance, tags, queue_blacklist)
 
-    def _yarn_cluster_metrics(self, rm_address, auth, ssl_verify, addl_tags):
+    def _yarn_cluster_metrics(self, rm_address, instance, addl_tags):
         """
         Get metrics related to YARN cluster
         """
-        metrics_json = self._rest_request_to_json(rm_address, auth, ssl_verify, YARN_CLUSTER_METRICS_PATH, addl_tags)
+        metrics_json = self._rest_request_to_json(rm_address, instance, YARN_CLUSTER_METRICS_PATH, addl_tags)
 
         if metrics_json:
 
@@ -277,12 +197,12 @@ class YarnCheck(AgentCheck):
             if yarn_metrics is not None:
                 self._set_yarn_metrics_from_json(addl_tags, yarn_metrics, YARN_CLUSTER_METRICS)
 
-    def _yarn_app_metrics(self, rm_address, auth, ssl_verify, app_tags, addl_tags):
+    def _yarn_app_metrics(self, rm_address, instance, app_tags, addl_tags):
         """
         Get metrics for running applications
         """
         metrics_json = self._rest_request_to_json(
-            rm_address, auth, ssl_verify, YARN_APPS_PATH, addl_tags, states=YARN_APPLICATION_STATES
+            rm_address, instance, YARN_APPS_PATH, addl_tags, states=YARN_APPLICATION_STATES
         )
 
         if metrics_json and metrics_json['apps'] is not None and metrics_json['apps']['app'] is not None:
@@ -302,11 +222,11 @@ class YarnCheck(AgentCheck):
 
                 self._set_yarn_metrics_from_json(tags, app_json, YARN_APP_METRICS)
 
-    def _yarn_node_metrics(self, rm_address, auth, ssl_verify, addl_tags):
+    def _yarn_node_metrics(self, rm_address, instance, addl_tags):
         """
         Get metrics related to YARN nodes
         """
-        metrics_json = self._rest_request_to_json(rm_address, auth, ssl_verify, YARN_NODES_PATH, addl_tags)
+        metrics_json = self._rest_request_to_json(rm_address, instance, YARN_NODES_PATH, addl_tags)
 
         if metrics_json and metrics_json['nodes'] is not None and metrics_json['nodes']['node'] is not None:
 
@@ -318,11 +238,11 @@ class YarnCheck(AgentCheck):
 
                 self._set_yarn_metrics_from_json(tags, node_json, YARN_NODE_METRICS)
 
-    def _yarn_scheduler_metrics(self, rm_address, auth, ssl_verify, addl_tags, queue_blacklist):
+    def _yarn_scheduler_metrics(self, rm_address, instance, addl_tags, queue_blacklist):
         """
         Get metrics from YARN scheduler
         """
-        metrics_json = self._rest_request_to_json(rm_address, auth, ssl_verify, YARN_SCHEDULER_PATH, addl_tags)
+        metrics_json = self._rest_request_to_json(rm_address, instance, YARN_SCHEDULER_PATH, addl_tags)
 
         try:
             metrics_json = metrics_json['scheduler']['schedulerInfo']
@@ -399,16 +319,12 @@ class YarnCheck(AgentCheck):
         else:
             self.log.error('Metric type "{}" unknown'.format(metric_type))
 
-    def _rest_request_to_json(self, address, auth, ssl_verify, object_path, tags, *args, **kwargs):
+    def _rest_request_to_json(self, url, instance, object_path, tags, *args, **kwargs):
         """
         Query the given URL and return the JSON response
         """
-        response_json = None
-
-        service_check_tags = ['url:{}'.format(self._get_url_base(address))] + tags
+        service_check_tags = ['url:{}'.format(self._get_url_base(url))] + tags
         service_check_tags = list(set(service_check_tags))
-
-        url = address
 
         if object_path:
             url = self._join_url_dir(url, object_path)
@@ -417,6 +333,33 @@ class YarnCheck(AgentCheck):
         if args:
             for directory in args:
                 url = self._join_url_dir(url, directory)
+
+        auth = None
+
+        # Authenticate our connection to JMX endpoint if required
+        kerberos = instance.get('kerberos')
+        username = instance.get('username')
+        password = instance.get('password')
+        if username is not None and password is not None:
+            auth = (username, password)
+        elif kerberos is not None:
+            if kerberos not in KERBEROS_STRATEGIES:
+                raise Exception('Invalid Kerberos strategy `{}`'.format(kerberos))
+
+            auth = requests_kerberos.HTTPKerberosAuth(
+                mutual_authentication=KERBEROS_STRATEGIES[kerberos],
+                delegate=is_affirmative(instance.get('kerberos_delegate', False)),
+                force_preemptive=is_affirmative(instance.get('kerberos_force_initiate', False)),
+                hostname_override=instance.get('kerberos_hostname'),
+                principal=instance.get('kerberos_principal')
+            )
+
+        ssl_verify = is_affirmative(instance.get('ssl_verify', True))
+
+        old_keytab_path = None
+        if 'kerberos_keytab' in instance:
+            old_keytab_path = os.getenv('KRB5_CLIENT_KTNAME')
+            os.environ['KRB5_CLIENT_KTNAME'] = instance['kerberos_keytab']
 
         self.log.debug('Attempting to connect to "{}"'.format(url))
 
@@ -460,7 +403,11 @@ class YarnCheck(AgentCheck):
                 message="Connection to {} was successful".format(url),
             )
 
-        return response_json
+            return response_json
+
+        finally:
+            if old_keytab_path is not None:
+                os.environ['KRB5_CLIENT_KTNAME'] = old_keytab_path
 
     def _join_url_dir(self, url, *args):
         """