+++ ibm_mq/CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## 1.0.1 / 2019-01-04
 
-* [Fixed] Fix Oldest Message Age. See [#2859](https://github.com/DataDog/integrations-core/pull/2859).
+* [Fixed] Fix Oldest Message Age. See [#2859][1].
 
 ## 1.0.0 / 2018-12-09
 
-* [Added] IBM MQ Integration. See [#2154](https://github.com/DataDog/integrations-core/pull/2154).
+* [Added] IBM MQ Integration. See [#2154][2].
 
+[1]: https://github.com/DataDog/integrations-core/pull/2859
+[2]: https://github.com/DataDog/integrations-core/pull/2154